pass remove-number numeric argument through to del_line()
FossilOrigin-Name: a0bd36c39edf211c0a19c228e02056522d7aa407

+++ src/features.c
@@ -1,6 +1,6 @@
 #include "features.h"
 
-char *del_item(unsigned int itemno) {
+char *del_item(uint16_t itemno) {
 	return del_line(itemno, QNAME);
 }
